+++ Utilities/Utilities/_Javascript/CosmoJavascript.swift
@@ -67,7 +67,7 @@ public final class CosmoJavascript: NSObject, SingletonProtocol {
         callNativeClient(functionName: "withdraw", params: [json], completion: completion)
     }
 
-    private func callNativeClient(functionName: String, params: [Any?], completion: @escaping JavascriptCompletion) {
+    public func callNativeClient(functionName: String, params: [Any?], completion: @escaping JavascriptCompletion) {
         loadV4Client { [weak self] _ in
             DispatchQueue.main.async {
                 if let runner = self?.v4ClientRunner {
@@ -115,59 +115,9 @@ public final class CosmoJavascript: NSObject, SingletonProtocol {
         }
     }
 
-    public func getMegavaulOwnerShares(payload: String) async -> String? {
-        return await call(functionName: "getMegavaultOwnerShares", params: [payload])
-    }
-
-    public func getMegavaultWithdrawalInfo(sharesToWithdraw: Double) async -> String? {
-        return await call(functionName: "getMegavaultWithdrawalInfo", params: [sharesToWithdraw])
-    }
-
-    public func depositToMegavault(subaccountNumber: Int32, amountUsdc: Double) async -> Result<ChainSuccessResponse, ChainError> {
-        let response = await call(functionName: "depositToMegavault", params: [subaccountNumber, amountUsdc])
-        return handleChainResponse(response)
-    }
-
-    public func withdrawFromMegavault(subaccountTo: String, shares: Double, minAmount: Double) async -> Result<ChainSuccessResponse, ChainError> {
-        let response = await call(functionName: "withdrawFromMegavault", params: [subaccountTo, shares, minAmount])
-        return handleChainResponse(response)
-    }
-
-    public func call(functionName: String, params: [Any?]) async -> String? {
-        return await withCheckedContinuation { continuation in
-            self.callNativeClient(functionName: functionName, params: params) { result in
-                continuation.resume(returning: result as? String)
-            }
-        }
-    }
-
     public func call(functionName: String, paramsInJson: String?, completion: @escaping JavascriptCompletion) {
         callNativeClient(functionName: functionName, params: paramsInJson != nil ? [paramsInJson!] : []) { result in
             completion(result)
-        }
-    }
-
-    // Helper function for parsing the response
-    private func handleChainResponse(_ response: String?) -> Result<ChainSuccessResponse, ChainError> {
-        guard let response = response else {
-            return .failure(.unknownError)
-        }
-
-        guard let jsonData = response.data(using: .utf8) else {
-            return .failure(.unknownError)
-        }
-
-        do {
-            // Try to decode the error first
-            if let error = try? JSONDecoder().decode(ChainErrorResponse.self, from: jsonData) {
-                return .failure(error.error)
-            }
-
-            // If no error, decode the success response
-            let success = try JSONDecoder().decode(ChainSuccessResponse.self, from: jsonData)
-            return .success(success)
-        } catch {
-            return .failure(.unknownError)
         }
     }
 }
